backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
httpx
pydantic[email]
pydantic-settings
resend
sendgrid
python-multipart
python-jose[cryptography]
PyJWT
passlib
bcrypt
